ffid and rid snowflakes.
Each FFID and RID is now generated with a special snowflake algorithm, as to make it less likely that a "clash error" (when an assigned FFID on one side of the bridge was already assigned for an object on the other) will occur in applications that run for a substantially long time.

Each snowflake is no more than 52 bits, to ensure that each snowflake is within Javascript's MAX_SAFE_INTEGER range.